Remove emptyss

+++ study-calendar-app/src/main/java/com/sbnz/studycalendarapp/service/StudyCalendarService.java
@@ -184,15 +184,15 @@ public class StudyCalendarService {
 		//po tom rasporedu- krenemo
 		//rasporedimo sesije ne manje od 2 sata redom svaki dan, i zaokruzene na vise od onog proseka
 		for(Obligation o: obs) {
-			int avgHrsPerDay = o.getStudyHours()/ (o.getStudyStartDate().datesUntil(o.getStudyEndDate()).collect(Collectors.toList())).size();
-			int hoursPerDay = avgHrsPerDay;
+			float avgHrsPerDay = o.getStudyHours()/ (o.getStudyStartDate().datesUntil(o.getStudyEndDate()).collect(Collectors.toList())).size();
+			float hoursPerDay = avgHrsPerDay;
 			if(avgHrsPerDay < 2) {
 				hoursPerDay = 2;
 			}
 			
-			int current = o.getStudyHours();
+			float current = o.getStudyHours();
 			for(LocalDate date : o.getStudyStartDate().datesUntil(o.getStudyEndDate()).collect(Collectors.toList())) {
-				if (current<= 0) {
+				if (current<= 0.0) {
 					sess = clearEmptySessions(sess,o);
 				}
 				//provera da li je taj dan zauzet?
